requirements.txt
aiogram
python-dotenv
httpx

